Clear conversion wanrings under MSVC

+++ gzip.cpp
@@ -40,7 +40,8 @@ void Gzip::WritePrestreamHeader()
 	AttachedTransformation()->Put(DEFLATED);
 	AttachedTransformation()->Put((byte)flags);		// general flag
 	AttachedTransformation()->PutWord32(m_filetime, LITTLE_ENDIAN_ORDER);	// time stamp
-	byte extra = (GetDeflateLevel() == 1) ? FAST : ((GetDeflateLevel() == 9) ? SLOW : 0);
+	byte extra = static_cast<byte>((GetDeflateLevel() == 1) ?
+	                FAST : ((GetDeflateLevel() == 9) ? SLOW : 0));
 	AttachedTransformation()->Put(extra);
 	AttachedTransformation()->Put(GZIP_OS_CODE);
